requirements.txt
streamlit
pillow
bcrypt
psycopg2-binary
